no more '~~' smartmatch operator in 5.41.3
Smartmatch was considered deprecated and dropped a
while ago. However, the corresponding operator '~~'
was still available (though it triggered the deprecation
warning).

With Perl version 5.41.3 the operator is finally gone,
so we can't overload it anymore.

Still providing it for the Perl versions it was available,
i.e., between 5.10 an 5.41.2 so that newer Data::DPath
doesn't break other code unneccessarily.